scripts: Add script for generating dbus interface xml files
Not all projects provide dbus interface files, and we're generally
only interested in a subset of the interfaces provided anyway,
so this script provides a way to pull down a subset of the dbus
interface for a service through introspection.

It's supposed to be run manually when a new proxy is needed in
the code and the result should get added to git.