input-sources-manager: Expose over user bus
Right now the only way to set the available keyboard layouts is from
GSettings or localed.

But some applications may need to be able to set the keyboard layout
independently over those.

This commit adds a new org.gnome.Kiosk.InputSources dbus interface that can
be used for this purpose.