build: Generate proxy for org.freedesktop.SessionManager
At the moment, GNOME Kiosk doesn't implement its own IBus UI.

This means we need to set GTK_IM_MODULE=ibus so we can leverage
the GTK3 based IBus UI. But we don't need to just set it for
the compositor, we need to set it for the whole session.

GNOME Session provides a D-Bus api for doing just that.

This commit makes a first step toward using that API by generating
a proxy to talk to gnome-session.

A future commit will make use of the proxy.