build: Generate proxy for org.gnome.DisplayManager.Manager
At the moment we manually call g_dbus_connection_call_sync to register
session.

I think the code would read a little nicer if it used a generated
proxy instead.

As a first step, toward that goal this commit generates a proxy.

A subsequent commit will change the code to use the proxy.