compositor: Add input source manager object
At the moment we're failing to initialize a keymap from system settings.
On Xorg this is probably okayish, since the keymap can come from xorg.conf
and the Xorg display server, but with the wayland native compositor, we
are the display server and we don't have a a config file, so we really
need some way to configure what the keymap is at startup. In a kiosk
situation, drawing this from system settings like, in this case, localed
makes sense.

This commit adds a input source manager object to manage this. This object
will eventually grow setters for changing the active layout at runtime,
which we'll expose over the bus for kiosk applications to leverage.

It's called "KioskInputSourcesManager" and not
"KioskKeyboardLayoutManager" or something, because it's going to
eventually grow support for IBus engines as well.