build: Add new gobject utilities
One thing we're going to need soon is the ability to dispatch a function
as the result of a flurry of events happening.  Any of the events should
lead to the function getting called, but we don't want to run it more
than once when several events happen at roughly the same time.

This commit adds a utility function for that purpose. Upon getting
called, it waits a 250ms coalescing interval before invoking the passed
in callback.  If the utility function is invoked during the coalescing
interval of a prior call, the function call is ignored.

For convenience, the callback function has two data arguments, the
first is the object associated with the call, and the second is
arbitrary.