Initial tail commit
This commit starts off the GNOME Kiosk project with a lone .gitignore
file. The content of it is rather ad hoc at the moment.  It's mainly
serving as a placeholder to allow the initial commit to be made.

This commit is very bare bones because the initial commit is kind of
a pain to update when doing interactive rebases and the like.

A subsequent commit will add the initial guts of the Kiosk project.