input-engine-manager: Use "us" layout by default
Certain input engines (like pinyin) set the XKB layout
to "default" which isn't a valid layout. It doesn't
matter that much, since input is getting overridden
anyway, but it does cause error messages to get
sent to the debug log.

This commit forces the "us" layout to get used in those
cases.