gobject-utils: Log when executing deferred tasks
At the moment, the code defers execution until "later" in
various parts of the code to ensure a flood of related events
doesn't lead to a flood of duplicated work.

But, its on the called code to log at the moment.

This commit adds logging to the generic part of the code to
for clarity.