gobject-utils: Add function for doing one shot call next iteration of event loop
At the moment it's a little cumbersome to just run an object method the
next time the event loop iterates. idle functions require having a glue
callback to make sure the source removed, and no existing API pays
attention to the lifetime of the object or its cancellable.

This commit adds a new utility:

kiosk_gobject_utils_queue_immediate_callback

That just runs the passed in method next time the event loop hits.