compositor: Support reading background from GSettings
While the solid gray color background is good enough for
many kiosk deployments, because many kiosk deployments have
a single full screen window, there are some users that
still want to be able to set a background image.

This commit adds a no frills implementation for that.

It does take special care to only set the background color
after the background image is loaded, so we don't see a brief
flicker of blue or whatever while the image is loaded from
disk.

Closes: https://gitlab.gnome.org/GNOME/gnome-kiosk/-/issues/5