check-format: Use git ls-files instead of find
When figuring out which source files to uncrustify, the check-format
script currently uses the find command to find all files that end in
.c or .h.  This isn't ideal, because it means stray files sitting
in the working tree that aren't tracked by git are getting processed
too.

This commit switches the script to use git ls-files instead, to avoid
that problem.