compositor: Default to plain background
commit 29e16b4ad8fa2066904fff274b2a7f7439c0e43f gave gnome-kiosk
a way to configure the background image.  While this is a useful
feature in some cases, defaulting to the distro provided background
is probably not what most users of gnome-kiosk are expecting.

This commit maintains the feature, but changes the default, so
out of the box we get the old behavior of a plain background.

It does this by creating a custom dconf profile for gnome-kiosk
that loads the background changes under the user settings, and
configures gnome-kiosk to be executed with DCONF_PROFILE=gnome-kiosk