Allow setting locale information from locale1
Apply system keyboard layout configuration when there is no overriding
configuration in place.

The code to do it was already in place but, due to bugs, it fails to
apply it in some cases and applies it erroneously in others.

This commit adds better bookkeeping to track which configuration source
input configuration comes from to apply the correct changes when
necessary.

Resolves: https://gitlab.gnome.org/GNOME/gnome-kiosk/-/issues/24