kiosk-window-tracker: Add KioskWindowTracker class
The purpose of the KioskWindowTracker is to connect the KioskApp and
KioskAppSystem together by maintaining the association between the
windows and the applications.

This code is a simplified and expunged version based on GNOME Shell
implementation of ShellWindowTracker.

This is preparation work for implementing support for the Shell
introspection DBus API.