shell-introspect: Add Shell Introspection support
Add support for the org.gnome.Shell.Introspect DBus API which is
required by Ponytail or even the desktop portals.

Support for the introspection API is enabled by default for the GNOME
and GTK desktop portals automatically.

To have access to the introspection API from all other clients,
gnome-kiosk needs to be started in unsafe mode (using the command line
option "--unsafe-mode").