compositor: Fix crash on VT change
When changing VTs, the monitors change andi, as a result, the background
image setting is reapplied.

However, when the image is already loaded, the data passed to the
callback function is of the wrong type, causing a crash in the callback
function.

That does not occur at startup because the background image is not
loaded yet.

Pass the right object to the callback function when the background image
is already loaded to avoid a crash of gnome-kiosk.

Closes: https://gitlab.gnome.org/GNOME/gnome-kiosk/-/issues/36
Fixes: 29e16b4a - compositor: Support reading background from GSettings