kiosk-app: Do not add the window in kiosk_app_new_for_window()
When a window has no matching application (like, when there is no
corresponding desktop file), a KioskApp is created for the window and
the window would be added to the newly-created KioskApp.

However, by doing so, we add the window before the calling window
tracker had a change to setup the state notify handler.

As a result, the window tracker is never notified of the state change
and the KioskApp is never added to the list of running applications in
the KioskAppState.

Too bad, since the introspection iterates on the running applications to
report the windows.

To avoid the issue, simply do not add the window so soon in
kiosk_app_new_for_window(), it will be added in time in track_window()
and the notifications will be correctly triggered.

Fixes: commit 50ae635dc - kiosk-app: Add a new KioskApp class