compositor: Make set-above configurable
GNOME Kiosk has heuristics to place windows on the top group.

However, in some cases, it may desirable to avoid that behavior as that
prevents the windows stacking from being changed by the user.

Add a new option "set-above" to the configuration file to control that
behaviour.

If not specified, the existing heuristics will be applied, not breaking
possible existing setups.

Closes: https://gitlab.gnome.org/GNOME/gnome-kiosk/-/issues/26