input-selector: Do not segfault if the DBus interface is not available
The DBus interface used is specific to GNOME Kiosk, and input-selector
assumes the DBus interface is available.

If ones tries to run the input selector on another Wayland compositor,
it will segfault trying to access a NULL object.

Let's be nice and log a critical message instead, and not segfault.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/60>