input-sources-manager: Do not crash if there is no X11Layouts
When queried for the X11Layout property, localed will try to get the
configuration from /etc/X11/xorg.conf.d/00-keyboard.conf.

If for some reason (the system was misconfigured or the installation was
altered), that directory is not accessible or not present, localed will
return a NULL string instead of the actual X11 layouts or even an empty
string if no X11 layout was ever configured.

In such a case, gnome-kiosk will segfault.

To avoid that potential problem, check for the X11 layouts to be not
NULL or return early if that's not the case.

Fixes: 38407233 - input-sources-manager: Add APIs for setting keymap
Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/68>