doc: Specify the right configuration file name
The CONFIG.md file mentions the file "gnome-kiosk.ini" as the name for
the window configuration file in multiple places.

This is just plain wrong and confusing, the actual name for the window
configuration file is "window-config.ini".

Specify the correct file name in the documentation.

Fixes: 9cb4f91724 - compositor: Add a window configuration file
Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/69>