search-app: Distinguish between Wayland and X11 session
The GNOME Kiosk Search Appliance session is identical between Wayland
and X11, making it indistinguishable when both Xorg and Wayland are
enabled.

That's a problem because users cannot tell which session uses X11 and
which one uses Wayland in the GDM session selection menu.

To avoid that issue, use two different desktop files, one for X11 and
another one for Wayland, and explicitly mention X11 and Wayland in the
name of the session to make it easy to identify each session in the GDM
session selection menu.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/73>