search-app: Keep the original session name
Previous change to differentiate between the X11 and Wayland sessions
changed the name of the desktop file, which can be disruptive for users
who had their session of choice selected before in gdm.

To avoid the problem when upgrading, let's revert to the original name
for the session file, but use different sub-directories instead.

Fixes: 0b57d4a6e - search-app: Distinguish between Wayland and X11 session
Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/74>