window-config: Account for non-existing window value
The way the window match logic works in the configuration is by using
keys to filter out sections.

If the client does not provide any value for the matching key, GNOME
Kiosk would assume a match.

That's a problem though, for example with window tags, as this means
that a window without any window tag would match a section specifying a
window tag.

This is also true for the other match values, of course, but that
remained unnoticed because most Wayland clients provide a value for the
other matching keys.

So instead of assuming no value means a match, assume no value is an
empty value, so that it can still match the keys even when there is a
wildcard.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/85>