compositor: Free objects on disposal
The Kiosk compositor would keep reference of its own objects and not
freeing them, causing those to be leaked on shutdown.

While a leak on shutdown is no big deal, as the memory allocated by the
process is to be returned to the system eventually anyway, it's good
practise to clean up the data we allocated.

Make sure to unref the objects the compositor allocated.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/87>