compositor: Allow inhibit shortcuts
By default, mutter will allow shortcuts to be inhibited, unless the
compositor implements an "inhibit shortcut" dialog to let the user
control that.

So, if we want to just allow shortcuts to be inhibited by default, all we
have to do is to remove our own function handler that does nothing and
let the default handler from mutter deal with the shortcut inhibition.

Closes: https://gitlab.gnome.org/GNOME/gnome-kiosk/-/issues/44
Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/90>