window-config: Add "lock-on-monitor" option
This new option is intended to lock a windows on a specific monitor,
preventing the windows from being moved or resized across different
monitors.

The external constraint mechanism from mutter is used to setup a
constraint on the window to keep the window within the designated
monitor.

Additionally, when windows are locked on a monitor, if the monitor is
removed, the windows are automatically hidden.

And if/when the monitor becomes available again, the windows which were
locked on that monitor are moved back to it and made visible again.

That ensures that the windows locked on a monitor do not accidentally
migrate to another monitor automatically if the designated monitor
vanishes.

In essence, this is somehow similar to "Zaphod" mode in Xorg.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/75>