window-config: Optimize window on monitor
The current implementation searches the entire configuration file to
determine if a window is set on a specific monitor.

That works for initial placement, yet if we are to repeatedly check
for that option, that might prove costly, searching constantly for
the connector name we already checked.

To optimize that search, store the windows set to be on a specific
monitor, along with the connector name, in a hash table and search that
hash table to find the specified output name.

This is still preparation work for the next commit, no functional change
intended.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/75>