window-config: Account for monitor set but not found
For now, if a monitor is specified but not found, it the same as if no
monitor was set.

Yet, if we want to lock a window to a given monitor, we need to
distinguish between no monitor being specified and a monitor being
specified but not found (like for example, after the monitor has
vanished).

No functional change at this point.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/75>