window-config: Add "lock-on-monitor-area" option
That allows to restrict a set of windows on to a specific area on screen.

This is similar to "lock-on-monitor" but restricted to a specified area
on the specified monitor.

Windows locked on a specific area cannot be moved or resized outside of
the specified area.

The external constraint mechanism from mutter is used to enforce the
constraint on the window.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/75>