magnifier: Add a magnifier for accessibility
Add the ability to zoom in the screen for accessibility purpose.

The magnifier in GNOME Kiosk uses the same settings as the rest of the
GNOME platform for setting the magnifying factor and enabling /
disabling the screen magnifier, namely:

 * org.gnome.desktop.a11y.applications screen-magnifier-enabled
 * org.gnome.desktop.a11y.magnifier mag-factor

That allows the screen magnifier to be enabled using the keyboard
shortcuts (defaults to <Alt><Super>[8]) and the zoom factor controlled
with the regular keyboard shortcuts (defaults to <Alt><Super>[=] to zoom
in and <Alt><Super>[-] to zoom out) with the "mediakeys" plugin from the
GNOME settings daemon.

The screen magnifier can also be enabled from the GNOME Kiosk
accessibility panel.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/106>