backgrounds: Fix wallpaper with multi-monitor
When a background image is specified, GNOME kiosk would load the image
and apply it of each monitor.

Unfortunately that code was not working correctly and would leave all
but one monitor with a background image visible.

Rework the code to use mutter's own API meta_background_set_file()
which takes care of loading the image and adjust it according to the
specified background style (stretched, centered, zoom, etc.)

So, not only that fixes the background image not showing on all
monitors, but it also removes a bunch of code which is not needed
anymore and also gives better results matching the settings in GNOME.

So, what's not to like?

Part-of: <https://gitlab.gnome.org/GNOME/gnome-kiosk/-/merge_requests/107>